Re-Add avaje-api-service and @Service to JsonbExtension
This is required in order to support the use of avaje-spi-serivce `@ServiceProvider` with
other custom JsonbExtension service implementations. 

Custom JsonbExtension implementations can be registered explicitly with META-INF/services
entries OR we can use `@ServiceProvider` for that to automatically happen.

The reason the `@Service` is required on the JsonbExtension interface is so that avaje-spi-service
actually ignores it correctly (and thus allowing avaje-jsonb-generator to process it with the 
generated JsonbExtension implementations). Without this  avaje-spi-service incorrectly 
assumes that the top level interface is the service interface and then doesn't ignore it and instead
generates an incorrect file in target/classes/META-INF/services.

---------

Co-authored-by: Rob Bygrave <robin.bygrave@gmail.com>