Add explicit license information to packages' POM file
Having the license information in the POM file allows it to be self-contained. A glance to it is enough to determine licensing.

This is especially important for private artifact mirrors, that often do not allow closed-source, source-available or hard-copyleft libraries to be used. Without the license information on the POM file, some of these repositories will deny mirroring this library.